Edited fab_location -Instead of building a new page, an overlay is opened when "Search Location" of fab_location button is tapped. -Added textcontrollers into "Search Location" for passing of data.
